+++ docs/CHANGELOG.rst
@@ -22,6 +22,7 @@ Changed
 * Improve variant table name in amplicon report
 * Prepend ``api/`` to all URL patterns in the Django test project
 * Set hisat2 process resources (memory) to 16GB
+* Filter LoFreq output VCF files to remove overlapping indels
 
 Fixed
 -----